Move BackupCheckEnum into a new util package
