+++ src/main/java/co/uk/mommyheather/advancedbackups/core/backups/BackupWrapper.java
@@ -16,6 +16,7 @@ import co.uk.mommyheather.advancedbackups.core.ABCore;
 import co.uk.mommyheather.advancedbackups.core.backups.gson.BackupManifest;
 import co.uk.mommyheather.advancedbackups.core.backups.gson.DifferentialManifest;
 import co.uk.mommyheather.advancedbackups.core.config.ConfigManager;
+import co.uk.mommyheather.advancedbackups.core.util.BackupCheckEnum;
 
 public class BackupWrapper {
